Add initial setup and testing framework
- Create init.sql to enable uuid-ossp extension.
- Add load_data.py for loading Titanic dataset into the database.
- Configure pytest with pytest.ini for testing setup and coverage.
- Initialize tests directory with __init__.py.
- Implement test cases in test_app.py for API endpoints and data models.

+++ src/models/person.py
@@ -98,7 +98,7 @@ class PersonSchema(Schema):
     passengerClass = fields.Int(required=True)
     name = fields.String(required=True)
     sex = fields.String(required=True)
-    age = fields.Int(required=True)
+    age = fields.Float(required=True)
     siblingsOrSpousesAboard = fields.Int(required=True)
     parentsOrChildrenAboard = fields.Int(required=True)
     fare = fields.Float(required=True)
